+++ R/test_ejam.R
@@ -316,6 +316,7 @@ x <- test_ejam(
         "test-calc_bg_extra_indicators.R",
         "test-calc_ejscreen_dataset.R",
         "test-calc_ejscreen_pctile_lookup_export.R",
+        "test-ejscreen-reference-adjustments.R",
         "test-datasets_arrow_publish.R",
         "test-ejscreen-stats.R",
         "test-ejscreen-pipeline-validate-vs-prior.R",